final project complete

+++ service/user_service.py
@@ -2,6 +2,7 @@ from typing import Optional, List
 
 from fastapi import HTTPException
 
+from api.internalApi import poll_service_api
 from model.user import User
 from repository import user_repository
 
@@ -31,5 +32,8 @@ async def register_user_by_id(user_id: int):
 
 
 async def delete_user_by_id(user_id):
+    answers_deleted = await poll_service_api.delete_answers_by_user_id(user_id)
+    if not answers_deleted:
+        raise HTTPException(status_code=500, detail=f"Failed to delete answers for user_id: {user_id}.")
     await user_repository.delete_user_by_id(user_id)
